feat(pm): expose embeddable command library
